merged

+++ inc/MiniRt.h
@@ -21,7 +21,7 @@ typedef struct	s_vect	t_vect;
 typedef struct s_ambient
 {
 	float	ratio;
-	int		color[3];
+	float	color[3];
 }			t_ambient;
 
 typedef struct	s_camera
@@ -36,15 +36,15 @@ typedef struct	s_light
 {
 	float	coord[3];
 	float	ratio;
-	int		color[3];
+	float	color[3];
 }			t_light;
 
 typedef struct s_sphere
 {
-	float	coord[3];
+	float	coord[3];	// change t_vect *center;
 	float	diametr;
 	float	radius;
-	int		color[3];
+	float	color[3];
 	struct s_sphere	*next;
 }			t_sphere;
 
@@ -52,7 +52,7 @@ typedef struct	s_plane
 {
 	float	coord[3];
 	float	orient[3];
-	int		color[3];
+	float	color[3];
 	struct s_plane	*next;
 }			t_plane;
 
@@ -60,7 +60,7 @@ typedef struct s_cylinder
 {
 	float	coord[3];
 	float	orient[3];
-	int		color[3];
+	float	color[3];
 	float	diametr;
 	float	height;
 	struct	s_cylinder *next;
